Make use of `aiida-shell==0.2.0`
The newly released `aiida-shell==0.2.0` allows to pass other data types
besides the `SinglefileData`, such as the `Int` and `Float` types. This
allows us to maintain the provenance of the integer and float values
that are parsed from the various scripts' output.

The example use case implementation is updated as well as the
documentation.